update: после протяжки бегут новобранцы

+++ Assets/Scripts/GameEntities/Interfaces/IDamagable.cs
@@ -1,5 +1,5 @@
 ﻿public interface IDamagable
 {
-    void TakeDamage(float damage , DamageSource damageSource);
+    void TakeDamage(int damage , DamageSource damageSource);
 }
